+++ Sources/AjarCore/Edit/CoreEditCommands.swift
@@ -52,7 +52,7 @@ extension EditReducer {
             return try applyTransformClipCommand(command, to: project)
         case .addClipMask, .removeClipMask, .moveClipMask, .setClipMask:
             return try applyClipMaskCommand(command, to: project)
-        case .setClipAudioMix, .clearClipAudioMix:
+        case .setClipAudioMix, .clearClipAudioMix, .setClipAudioRetimeMode:
             return try applyClipAudioMixCommand(command, to: project)
         case .setClipAudioCrossfade, .removeClipAudioCrossfade:
             return try applyClipAudioCrossfadeCommand(command, to: project)